+++ docs/README.md
@@ -14,4 +14,4 @@ the moment. See https://github.com/asciidoctor/asciidoctor/issues/1907
 
 Apart from that, this folder looks messier than it is -- what you are probably
 looking for are the `manual.html`, `manual.pdf` and possibly the `WORDLIST.md`
-files. There are also the listing files here generated by the Ophis assembler.
+files. There are also the listing files here generated by the assembler.
