+++ docs/manual.html
@@ -4,7 +4,7 @@
 <meta charset="UTF-8">
 <meta http-equiv="X-UA-Compatible" content="IE=edge">
 <meta name="viewport" content="width=device-width, initial-scale=1.0">
-<meta name="generator" content="Asciidoctor 2.0.20">
+<meta name="generator" content="Asciidoctor 2.0.23">
 <meta name="keywords" content="forth, 6502, assembler, programming, 8-bit, vintage, retro">
 <meta name="author" content="Sam Colwell, Scot W. Stevenson, Patrick Surry">
 <title>Manual for Tali Forth 2 for the 65c02</title>
@@ -142,7 +142,7 @@ p a>code:hover{color:rgba(0,0,0,.9)}
 #content::before{content:none}
 #header>h1:first-child{color:rgba(0,0,0,.85);margin-top:2.25rem;margin-bottom:0}
 #header>h1:first-child+#toc{margin-top:8px;border-top:1px solid #dddddf}
-#header>h1:only-child,body.toc2 #header>h1:nth-last-child(2){border-bottom:1px solid #dddddf;padding-bottom:8px}
+#header>h1:only-child{border-bottom:1px solid #dddddf;padding-bottom:8px}
 #header .details{border-bottom:1px solid #dddddf;line-height:1.45;padding-top:.25em;padding-bottom:.25em;padding-left:.25em;color:rgba(0,0,0,.6);display:flex;flex-flow:row wrap}
 #header .details span:first-child{margin-left:-.125em}
 #header .details span.email a{color:rgba(0,0,0,.85)}
@@ -164,6 +164,7 @@ p a>code:hover{color:rgba(0,0,0,.9)}
 #toctitle{color:#7a2518;font-size:1.2em}
 @media screen and (min-width:768px){#toctitle{font-size:1.375em}
 body.toc2{padding-left:15em;padding-right:0}
+body.toc2 #header>h1:nth-last-child(2){border-bottom:1px solid #dddddf;padding-bottom:8px}
 #toc.toc2{margin-top:0!important;background:#f8f8f7;position:fixed;width:15em;left:0;top:0;border-right:1px solid #e7e7e9;border-top-width:0!important;border-bottom-width:0!important;z-index:1000;padding:1.25em 1em;height:100%;overflow:auto}
 #toc.toc2 #toctitle{margin-top:0;margin-bottom:.8rem;font-size:1.2em}
 #toc.toc2>ul{font-size:.9em;margin-bottom:0}
@@ -329,7 +330,7 @@ a.image{text-decoration:none;display:inline-block}
 a.image object{pointer-events:none}
 sup.footnote,sup.footnoteref{font-size:.875em;position:static;vertical-align:super}
 sup.footnote a,sup.footnoteref a{text-decoration:none}
-sup.footnote a:active,sup.footnoteref a:active{text-decoration:underline}
+sup.footnote a:active,sup.footnoteref a:active,#footnotes .footnote a:first-of-type:active{text-decoration:underline}
 #footnotes{padding-top:.75em;padding-bottom:.75em;margin-bottom:.625em}
 #footnotes hr{width:20%;min-width:6.25em;margin:-.25em 0 .75em;border-width:1px 0 0}
 #footnotes .footnote{padding:0 .375em 0 .225em;line-height:1.3334;font-size:.875em;margin-left:1.2em;margin-bottom:.2em}
@@ -1072,9 +1073,9 @@ Type 'bye' to exit</pre>
 <div class="paragraph">
 <p>Because these are the last high-level commands Tali Forth executes, this
 functions as a primitive self-test. If you have modified the high level Forth
-words in either <code>forth_words.fs</code> or <code>user_words.fs</code>, the boot process might
+words in your platform&#8217;s <code>platform_forth.fs</code> the boot process might
 fail with a variant of the error message "unknown word". The built-in, native
-words should always work. For this <code>dump</code>  is a built-in word&#8201;&#8212;&#8201;it is very
+words should always work. The <code>dump</code>  built-in word is very
 useful for testing.</p>
 </div>
 </div>
@@ -5501,10 +5502,10 @@ to find that word.</p>
 <h4 id="_structure_of_the_header_list">Structure of the Header List</h4>
 <div class="paragraph">
 <p>Tali Forth distinguishes between three different word sources: The <strong>native
-words</strong> that are hard-coded in the <code>words/*.asm</code> files, the <strong>Forth words</strong>
-from <code>forth_words.asm</code> which are defined as high-level words and then generated
-at run-time when Tali Forth starts up, and <strong>user words</strong> in the file
-<code>user_words.asm</code>.</p>
+words</strong> that are hard-coded in the <code>words/*.asm</code> files,
+any <strong>native words</strong> specific to the platform in <code>platform/xxx/platform_words.asm</code>,
+and any <strong>Forth words</strong> defined for the platform in <code>platform/xxx/platform_forth.fs</code>,
+which are compiled from source at run-time when Tali Forth starts up.</p>
 </div>
 <div class="paragraph">
 <p>Tali has an unusually high number of native words in an attempt to make the
@@ -5664,10 +5665,9 @@ buffer.</p>
 </div>
 </div>
 <div class="paragraph">
-<p>Tali Forth uses <code>evaluate</code> to load high-level Forth words from the file
-<code>forth_words.fs</code> and, if present, any extra, user-defined words from
-<code>user_words.fs</code>.  The code in these files has all comments removed and all
-whitespace replaced with a single space between words.  This minimized version
+<p>Tali Forth uses <code>evaluate</code> to load high-level Forth words from the platform&#8217;s
+<code>platform_forth.fs</code>, if present.  Forth code in these files has all comments removed and all
+whitespace replaced with a single space between words.  This compact version
 is assembled directly into the ROM image as a string that will be evaluated at
 startup.</p>
 </div>
@@ -6612,7 +6612,7 @@ a Tali Forth version.</p>
 </li>
 <li>
 <p>Write the word in Forth in the interpreter. After it has been tested
-interactively, add a high-level version to the file <code>forth_code/forth_words.fs</code>.</p>
+interactively, add a high-level version to your 'platform_forth.fs` file.</p>
 </li>
 <li>
 <p>Add automatic tests for the new word to the test suite. Ideally, there will
@@ -10739,7 +10739,7 @@ under <a href="https://www.ubuntu.com/">Ubuntu</a> Linux 16.04 LTS.</p>
 <div id="footer">
 <div id="footer-text">
 Version 1.1<br>
-Last updated 2025-06-24 19:35:50 -0400
+Last updated 2025-07-01 15:02:47 -0400
 </div>
 </div>
 </body>
