+++ examples/forth/vibe22tali.fth
