+++ examples/words/bind.asm
