+++ examples/words/block-ext.asm
