+++ examples/words/byte.asm
