+++ examples/words/core-ext.asm
