+++ examples/words/dmp.asm
