+++ examples/words/rand.asm
