+++ examples/words/string-ext.asm
