+++ words/all.asm
@@ -68,25 +68,23 @@ _load_user_vars_loop:
 
                 jsr w_cr
 
-                ; Define high-level words in forth_words.asc via EVALUATE,
-                ; followed by any user-defined words from user_words.asc.
-                ; These are stored sequentially in ROM so we can evaluate them together.
-                ; If you have neither, this section can be commented out.
+                ; EVALUATE any platform-provided forth code
+                ; If there is none, this section can be commented out.
                 dex
                 dex
                 dex
                 dex
 
                 ; start address goes NOS
-                lda #<forth_words_start
+                lda #<user_words_start
                 sta 2,x
-                lda #>forth_words_start
+                lda #>user_words_start
                 sta 3,x
 
                 ; length goes TOS; let the assembler do the math
-                lda #<(user_words_end-forth_words_start)
+                lda #<(user_words_end-user_words_start)
                 sta 0,x
-                lda #>(user_words_end-forth_words_start)
+                lda #>(user_words_end-user_words_start)
                 sta 1,x
 
                 jsr w_evaluate
